(PC-17542)[PRO] refactor: Refactoring `serializeApiErrors` to reuse it.
